buttons and movie widgets separated from quiz widget
