+++ lib/presentation/view/widgets/movie_item.dart
@@ -1,0 +1,34 @@
+import 'package:cached_network_image/cached_network_image.dart';
+import 'package:flutter/material.dart';
+
+class MovieItem extends StatelessWidget {
+  final String urlPreview;
+
+  const MovieItem({
+    super.key,
+    required this.urlPreview,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    return Card(
+      elevation: 10,
+      margin: const EdgeInsets.all(30),
+      shape: RoundedRectangleBorder(
+        borderRadius: BorderRadius.circular(24),
+      ),
+      child: SizedBox(
+        height: 500,
+        child: ClipRRect(
+          borderRadius: BorderRadius.circular(24),
+          child: CachedNetworkImage(
+            imageUrl: urlPreview,
+            placeholder: (context, url) => const Center(child: CircularProgressIndicator()),
+            errorWidget: (context, url, error) => const Icon(Icons.error),
+            fit: BoxFit.cover,
+          ),
+        ),
+      ),
+    );
+  }
+}
